feat: Add recharts library for data visualization
The code changes add the "recharts" library as a dependency in the package.json file. This library is used for data visualization in the application. The changes also include importing the necessary components from "recharts" and adding a bar chart to display user statistics. This enhances the application's ability to present data in a visually appealing and informative way.